[core] Pull annotation typedefs to own header; add AnnotationID
